Try to auto-deduce the types provided for ctx.run() (#63)
This commit try to deduce the type that is provided to the
ctx.run(), in the case where a `serde` parameter is not passed.
If type hints are there we will deduce the correct serializer from the type hints.
If the type hints are not there, we will consult the `type_hint`
parameter and try to deduce a serializer from there. If nothing worked,
we will default to the DefaultSerde()

This closes #32