Rename created Docker image into test-services-python
I was confused by our naming pattern of the Docker image. The image that we
are publishing is called test-services-python and the one we build in the
integration workflow is called python-test-services. When reading the workflow
run on GHA it took me quite some time to figure out where this image comes
from.