re-raise TerminalError/SdkBaseException that are accidentally wrapped (#151)
This commit helps in the situations where a TerminalError or the internal (SdkBaseException) was
accidentally wrapped with a try - except block and then re-thrown as the cause.
For example:
```
try:
      ctx.run_typed( ) # throws TerminalError
except Exception as e: # big no no
      raise RuntimeError( , cause=e)
```
Before that commit this will be considered as a retryable error, and with this commit
applied this will be identified as a TerminalError.