Support handler/run_typed Optional return types (#158)
This commit adds support for handler *return types* that are `Union[T, None].`
With this commit, the correct serializer will be detected (and the correct json schema) for the following situations

```python
async def greeter(ctx: Context, name: str) -> MyResponse | None:
         return MyResponse(..)
```

With this commit the correct serializer will be used (i.e. the serializer for MyResponse) instead of a generic JSON one.
This also applies for `ctx.run_typed` where the provided function had the return type of `T | None`.
There is no need to pass a `type_hint`in this situation.

Previously:

```
async def my_effect() -> PydanticMessage | None
     res = ctx.run_typed(... , type_hint=PydanticMessage)
     
```